+++ lib/src/ArgusV4l2Capture.cpp
@@ -830,11 +830,7 @@ int ArgusV4l2Capture::setManualAnalogGain(int percent)
     return -1;
 }
 
-int ArgusV4l2Capture::setManualAnalogGainReal(int db)
-{
-    return -1;
-}
-
+int ArgusV4l2Capture::setManualAnalogGainReal(float db) { return -1; }
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////////////////
 //////////////////////////////// DIGITAL GAIN (ISP) CONTROL/REQUEST ///////////////////////////////////////////
